Day 61 of 100
Today I fixed something that has annoyed me for the longest time. The progress indicator was in the scaffold, so whenever it was loading it'll show a blank screen with a progress indicator.

I fixed it by putting everything in a stack and progress indicator under the body in the stack.