Day 66 of 100
I found the bug. I was setting the ID as a hash of the habit title, so there could only be 1 notification per habit

Now I have updated the id to be  a hash of the habit+time+day.

So 1 habit for every time of every day.

In summary, there can be up to 21 notifications per habit. Each with an id gotten from a hash of the addition of the habit's title and the time and day of the notification.